fixed the input to take double instead of int and now you can handle all the suggested operations

+++ src/Main.java
@@ -4,13 +4,19 @@ public class Main {
     public static void main(String[] args) {
         Scanner s = new Scanner(System.in);
         System.out.print("Enter first number: ");
-        int firstNumber = s.nextInt();
+        double firstNumber = s.nextDouble();
         System.out.print("Enter second number: ");
-        int secondNumber = s.nextInt();
-        int sum = firstNumber + secondNumber;
-        int difference = firstNumber - secondNumber;
+        double secondNumber = s.nextDouble();
+        double sum = firstNumber + secondNumber;
+        double difference = firstNumber - secondNumber;
+
         double division = firstNumber / secondNumber;
         double remainder = firstNumber % secondNumber;
-        System.out.println("The difference is " + difference + " The sum is " + sum + "The result of division was " + division + " with a remainder of  " + remainder);
+
+            System.out.println("The difference is " + difference + " The sum is " + sum + "The result of division was " + division + " with a remainder of  " + remainder);
+
     }
 }
+
+
+//try {} catch {} don't use finally
